Settings: horizontally scrollable menu with overflow arrows on mobile
On phones the settings menu pills wrapped into a tall block that pushed
the content down. The row now rides the same horizontal scroller pattern
as the history filters: sideways scroll with the scrollbar hidden, and a
small arrow on each side shown only while entries remain past that edge.

The scroll bookkeeping (overflow state, wheel translation, arrow
rendering) is extracted from the history page into a shared
components/chips-scroll component used by both pages; each page keeps
its own styling.
