+++ front/src/routes/settings/style.css
@@ -16,17 +16,100 @@
 
 /* Section navigation: a horizontal row of frosted pills under the title —
    the dashboard switcher's grammar (active = inverted ink pill). With the
-   app nav in the left rail, a second vertical menu was redundant. The row
-   wraps instead of collapsing to icons: several entries share an icon, so
-   labels stay at every width. */
-.settingsTabs {
-  display: flex;
-  flex-wrap: wrap;
-  gap: 8px;
+   app nav in the left rail, a second vertical menu was redundant. Labels
+   stay at every width (several entries share an icon): instead of wrapping
+   into a tall block on phones, the row rides the shared Horizon chips
+   scroller (the history filters' pattern) — sideways scroll with the
+   scrollbar hidden, and an overflow arrow per side while pills remain
+   past that edge. */
+.settingsTabsWrapper {
+  --st-scroll-btn: 2rem;
+
+  position: relative;
   margin-bottom: 1.5rem;
 }
 
+.settingsTabs {
+  display: flex;
+  gap: 8px;
+  overflow-x: auto;
+  /* the pills float on the scene with soft drop shadows: bleed the scroll
+     box symmetrically so the shadows aren't sheared at its edges, and give
+     the room back so the wrapper keeps the bare pill-row height (which also
+     keeps the absolute arrows centered on the pills) */
+  padding: 18px 4px;
+  margin: -18px -4px;
+  -webkit-overflow-scrolling: touch;
+  scrollbar-width: none;
+  -ms-overflow-style: none;
+}
+
+.settingsTabs::-webkit-scrollbar {
+  display: none;
+}
+
+/* Reveal a thin scrollbar on hover/focus for mouse users (e.g. Windows) */
+.settingsTabs:hover,
+.settingsTabs:focus-within {
+  scrollbar-width: thin;
+}
+
+.settingsTabs:hover::-webkit-scrollbar,
+.settingsTabs:focus-within::-webkit-scrollbar {
+  display: block;
+  height: 4px;
+}
+
+.settingsTabs:hover::-webkit-scrollbar-thumb,
+.settingsTabs:focus-within::-webkit-scrollbar-thumb {
+  background: rgba(28, 35, 52, 0.2);
+  border-radius: 4px;
+}
+
+.tabsScrollBtn {
+  position: absolute;
+  top: 50%;
+  z-index: 2;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  width: var(--st-scroll-btn);
+  height: var(--st-scroll-btn);
+  padding: 0;
+  color: var(--gl-ink);
+  background: var(--gl-surface-strong);
+  border: 1px solid var(--gl-border);
+  border-radius: 50%;
+  box-shadow: var(--gl-shadow-soft);
+  cursor: pointer;
+  transform: translateY(-50%);
+  transition: color 0.15s ease;
+}
+
+.tabsScrollBtn:hover {
+  color: var(--gl-accent);
+}
+
+.tabsScrollBtnLeft {
+  left: 0;
+}
+
+.tabsScrollBtnRight {
+  right: 0;
+}
+
+/* keep the first/last pill clear of a visible overflow arrow (its width +
+   a small gap), so the padding follows the coarse-pointer size */
+.settingsTabsWrapper:has(.tabsScrollBtnLeft) .settingsTabs {
+  padding-left: calc(var(--st-scroll-btn) + 0.5rem);
+}
+
+.settingsTabsWrapper:has(.tabsScrollBtnRight) .settingsTabs {
+  padding-right: calc(var(--st-scroll-btn) + 0.5rem);
+}
+
 .tabLink {
+  flex-shrink: 0;
   display: inline-flex;
   align-items: center;
   gap: 0.45rem;
@@ -65,6 +148,12 @@
 @media (pointer: coarse) {
   .tabLink {
     height: 2.75rem;
+  }
+
+  /* the overflow arrows are real tap targets (they render whenever the
+     row overflows, touch included): same floor as the pills */
+  .settingsTabsWrapper {
+    --st-scroll-btn: 2.75rem;
   }
 }
